Tile impl Draw complications
Because neither Rect nor Option<Texture> derive Rustc*code, I've had to
move all that information outside of the Tile struct. This awkwardly
forced me to either have a bunch of static Texture's laying around or
add an &Option<Texture> to the Draw trait and do a teensy bit of the
Tile drawing logic in main.

I went with the latter. I may yet change back to the former.